requirements.txt
Django~=4.1.7
djangorestframework~=3.14.0
django-leaflet~=0.28.3
bokeh~=3.1.0
djangorestframework-gis~=1.0
django-multiselectfield~=0.1.12
django-filter~=22.1
django-map-widgets~=0.4.1
python-dotenv~=1.0.0